added full sync note
Some settings never getting a sync, this is what I complain about years. You need to manually create an entry for this in about:config to 'force' sync. Of course the opposite also must be work, there is no 'reset' button based on browser version because you only can reset everything and not only the core functions.